housekeeping

+++ ch03/ex04/surface.go
@@ -10,12 +10,12 @@ import (
 )
 
 var (
-	width, height = 600, 320            // a canvas size in pixels
-	cells         = 100                 // number of grid cells
-	xyRange       = 30.0                // axis range (-xyRange..+xyRange)
-	xyScale       = width / 2 / xyRange // pixels per x or y unit
-	zScale        = height * 0.4        // pixels per z unit
-	angle         = math.Pi / 6         // angle of x, y axes (=30°)
+	width, height = 600, 320                     // a canvas size in pixels
+	cells         = 100                          // number of grid cells
+	xyRange       = 30.0                         // axis range (-xyRange..+xyRange)
+	xyScale       = float64(width) / 2 / xyRange // pixels per x or y unit
+	zScale        = float64(height) * 0.4        // pixels per z unit
+	angle         = math.Pi / 6                  // angle of x, y axes (=30°)
 )
 
 var sin30, cos30 = math.Sin(angle), math.Cos(angle) // sin(30°), cos(30°)
@@ -47,15 +47,15 @@ func svg(w, h int) string {
 
 func corner(i, j int) (float64, float64) {
 	// Find point (x,y) at corner of cell (i,j).
-	x := xyRange * (float64(i)/cells - 0.5)
-	y := xyRange * (float64(j)/cells - 0.5)
+	x := xyRange * (float64(i)/float64(cells) - 0.5)
+	y := xyRange * (float64(j)/float64(cells) - 0.5)
 
 	// Compute surface height z.
 	z := f(x, y)
 
 	// Project (x,y,z) isometrically onto 2-D SVG canvas (sx, sy).
-	sx := width/2 + (x-y)*cos30*xyScale
-	sy := height/2 + (x+y)*sin30*xyScale - z*zScale
+	sx := float64(width/2) + (x-y)*cos30*xyScale
+	sy := float64(height/2) + (x+y)*sin30*xyScale - z*zScale
 	return sx, sy
 }
 
@@ -79,7 +79,7 @@ func colorIndex(i, j int) string {
 	// z: min = -0.21722891503668823, max = 0.9850673555377986 -> (0.0 to 1.0) -> 0 to 255
 	const minValue = 0.3
 	const maxValue = 1.0
-	z := f(xyRange*(float64(i)/cells-0.5), xyRange*(float64(j)/cells-0.5))
+	z := f(xyRange*(float64(i)/float64(cells)-0.5), xyRange*(float64(j)/float64(cells)-0.5))
 	r := (z + minValue) * (256 / (minValue + maxValue))
 	g := 0
 	b := 255 - r
